requirements.txt
streamlit
pandas
sentence-transformers
torch
google-generativeai
scikit-learn